Candidate Model Added for candidate side registration and login
