--- CandidateApp/urls.py
@@ -1,24 +1,0 @@
-from django.db import models
-
-class Candidate(models.Model):
-    name = models.CharField(max_length=255)
-    email = models.EmailField(unique=True)
-    address = models.TextField()
-    password = models.CharField(max_length=255)
-    refresh_token = models.CharField(max_length=255, blank=True, null=True)
-
-    def __str__(self):
-        return self.name
-
-    def set_password(self, raw_password):
-        from django.contrib.auth.hashers import make_password
-        self.password = make_password(raw_password)
-
-    def check_password(self, raw_password):
-        from django.contrib.auth.hashers import check_password
-        return check_password(raw_password, self.password)
-
-    def generate_refresh_token(self):
-        import uuid
-        self.refresh_token = str(uuid.uuid4())
-        self.save()
